[ldm] extracted file watching
